cnki/requirements.txt
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
Deprecated==1.2.13
filelock==3.6.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
msedge-selenium-tools==3.141.4
outcome==1.1.0
packaging==21.0
parsel==1.6.0
peewee==3.14.10
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyExecJS==1.5.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
queuelib==1.6.2
redis==4.2.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
scrapy-redis==0.7.2
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tldextract==3.2.0
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.2.0
typing_extensions==4.1.1
urllib3==1.26.9
w3lib==1.22.0
wrapt==1.14.0
wsproto==1.1.0
zipp==3.7.0
zope.interface==5.4.0
